Rename favorite track entity
