+++ app/src/main/java/data/db/dao/FavoriteTracksDao.kt
@@ -4,21 +4,21 @@ import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
-import data.db.entity.TrackEntity
+import data.db.entity.FavoriteTrackEntity
 
 @Dao
 interface FavoriteTracksDao {
 
     @Insert
-    suspend fun add(track: TrackEntity)
+    suspend fun add(track: FavoriteTrackEntity)
 
     @Query("SELECT * FROM favorite_tracks WHERE trackId = :id")
-    suspend fun getById(id: Int): TrackEntity?
+    suspend fun getById(id: Int): FavoriteTrackEntity?
 
     @Delete
-    suspend fun delete(track: TrackEntity)
+    suspend fun delete(track: FavoriteTrackEntity)
 
     @Query("SELECT * FROM favorite_tracks ORDER BY timestamp DESC")
-    suspend fun getAll(): List<TrackEntity>
+    suspend fun getAll(): List<FavoriteTrackEntity>
 
 }
