+++ app/src/main/java/data/db/entity/FavoriteTrackEntity.kt
@@ -4,7 +4,7 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "favorite_tracks")
-data class TrackEntity(
+data class FavoriteTrackEntity(
     @PrimaryKey
     val trackId: Int,
     val trackName: String,
