+++ app/src/main/java/di/DataModule.kt
@@ -1,7 +1,7 @@
 package di
 
 import androidx.room.Room
-import data.converters.TrackDbConvertor
+import data.convertors.FavoriteTrackDbConvertor
 import data.db.AppDatabase
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
@@ -16,6 +16,6 @@ val dataModule = module {
         ).build()
     }
 
-    factory { TrackDbConvertor() }
+    factory { FavoriteTrackDbConvertor() }
 
 }
